Local theme: modern-child
Start a child theme that will let us tweak templates and styles in the
upstream theme that we have selected. Initial use is integration with
the co-authors plugin so that posts can have multiple people in the
by-line. Chances are pretty good that we will find some other things to
override before we are done as well.